# ci: improve version module performance