Before:

```
$ time dagger call -m github.com/dagger/dagger/version@main version
v0.18.15-250728111339-1a8c04db77d6

Full trace at https://dagger.cloud/jedevc/traces/fa964659c3e20b2df74c3b43590d70ab
dagger-dev call -m github.com/dagger/dagger/version@main version  3.44s user 0.88s system 4% cpu 1:37.37 total

$ time dagger call -m github.com/dagger/dagger/version@main version
v0.18.15-250728111339-1a8c04db77d6

Full trace at https://dagger.cloud/jedevc/traces/8276c6e06d137a4fd0b665f431100f4e
dagger-dev call -m github.com/dagger/dagger/version@main version  0.75s user 0.21s system 9% cpu 10.175 total
```

After:

```
$ time dagger call -m github.com/jedevc/dagger/version@improve-version-perf version
v0.18.15-250728150307-c367a1eb271b

Full trace at https://dagger.cloud/jedevc/traces/45d5694213e5a322564c55867b766afd
dagger-dev call -m github.com/jedevc/dagger/version@improve-version-perf   3.42s user 0.79s system 4% cpu 1:36.15 total

$ time dagger call -m github.com/jedevc/dagger/version@improve-version-perf version
v0.18.15-250728150307-c367a1eb271b

Full trace at https://dagger.cloud/jedevc/traces/daa306ed18f07de4e1f3504f0273a439
dagger-dev call -m github.com/jedevc/dagger/version@improve-version-perf   0.44s user 0.12s system 10% cpu 5.276 total
```

The uncached performance is pretty much identical (we're still having to do all the same work), but the cache performance is twice as fast (from ~10s to ~5s) - we no longer need to pull the whole repo everytime.

|                | Uncached Total Time | Cached Total Time |
|----------------|---------------------|-------------------|
| [**Before**](https://dagger.cloud/jedevc/traces/8276c6e06d137a4fd0b665f431100f4e?span=394fb270e360d8c4)     | 97.37s              | 10.175s           |
| [**After**](https://dagger.cloud/jedevc/traces/daa306ed18f07de4e1f3504f0273a439?span=77a3d59c1f51993d)      | 96.15               | 5.276s            |