@jedevc 2025-07-31
:angry: hm, yeah, this is failing because `merge-base` can't run without a full history. This is kind of frustrating - because the default `depth` of the `.git` directory is 1.

I think we still need some git api changes - will brainstorm some. I think we want something like https://github.com/dagger/dagger/issues/10200.